models/Keyword Extraction_Disease Prediction/requirements.txt
streamlit
Pillow
docx2txt
pdfplumber
pytesseract 
tesseract
wikipedia
opencv-python-headless
scikit-learn
